requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.8.1
aniso8601==6.0.0
async-timeout==4.0.3
attrs==23.1.0
autopep8==1.7.0
Babel==2.10.3
black==22.8.0
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.3
distlib==0.3.5
docopt==0.6.2
et-xmlfile==1.1.0
filelock==3.7.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Migrate==3.1.0
Flask-Moment==1.0.4
Flask-WTF==1.0.1
frozenlist==1.4.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
Mako==1.2.1
MarkupSafe==2.1.1
multidict==6.0.4
mypy-extensions==0.4.3
openai==0.28.1
openpyxl==3.1.2
packaging==21.3
pathspec==0.9.0
Pillow==10.1.0
pipreqs==0.4.13
platformdirs==2.5.2
pycodestyle==2.9.1
pyparsing==3.0.9
PyPDF2==2.12.1
python-dateutil==2.8.2
python-docx==1.0.1
python-dotenv==0.20.0
pytz==2019.1
requests==2.31.0
rq==1.15.1
six==1.12.0
toml==0.10.2
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
virtualenv==20.15.1
Werkzeug==3.0.0
WTForms==3.0.1
yarg==0.1.9
yarl==1.9.2
